Implement sparse reward system and reward_mapping attribute
The following patch implements a single entrypoint for the reward
system as a dictionary in the reward_mapping attribute.
Based on whether the sparse_flag is true or not,
a dense reward or a sparse reward mapping can be defined.

This patch only defines the sparse reward system. Follow up
work on the dense reward to be implemented as part
of: https://github.com/sjsu-interconnect/ourhexgame/issues/6

Callers that used a hardcoded reward have also been updated
to look up an appropiate reward in the reward_mapping instead.