Return a copy of the board in the observe method
We are only returning the self.board using the observe method.

This is a bug in the environment. Why? Assume the agent holds the observation from the first step.
In the third step, the observation from the first step (held by the agent) will mutate
to reflect the state of the board at the third step. This is undesired behavior.

The fix will be to return a simple .copy version of the board in the observe method

Fixes: https://github.com/sjsu-interconnect/ourhexgame/issues/41