Files for group7 agent
The following patch adds the source code for the G07Agent.

I am also adding a script to install a couple of requirements
for the agent (just in case it needs to be done).

I had to re-write a couple of imports to match the folder
structure of this repo, but I tested the import
in ipython and it should be good to go.

TODO: Update the model files.